+++ util.h
@@ -2,7 +2,7 @@
 int timenum_parse(const char *str, time_t *res);
 
 /* ret==-1 : failure, ret<bufsize : success, bufsize<=ret : truncated */
-int format_gmtoff(char *buf, size_t bufsize, int gmtoff);
+int format_gmtoff(char *buf, size_t bufsize, long gmtoff);
 
 /* ret==-1 : failure, ret<bufsize : success, bufsize<=ret : truncated */
 int format_wday(char *buf, size_t bufsize, int wday);
